fix(rbac): add team:read to member role and use team biz_key in permissions map
- member preset role was missing team:read, causing 403 on team detail page
- GetUserTeamPermissions now returns map[int64][]string keyed by team biz_key
  instead of internal id, so permission checks work correctly with biz_key-based
  team context
- Update all callers, tests, and integration tests accordingly

+++ backend/internal/handler/permission_handler_test.go
@@ -149,7 +149,7 @@ func TestPermissionHandler_GetUserPermissions_Success(t *testing.T) {
 		getUserPermissionsFn: func(_ context.Context, userID uint) (*service.UserPermissions, error) {
 			return &service.UserPermissions{
 				IsSuperAdmin:    false,
-				TeamPermissions: map[uint][]string{10: {"main_item:create", "main_item:read"}},
+				TeamPermissions: map[int64][]string{10: {"main_item:create", "main_item:read"}},
 			}, nil
 		},
 	}
@@ -186,7 +186,7 @@ func TestPermissionHandler_GetUserPermissions_SuperAdmin(t *testing.T) {
 		getUserPermissionsFn: func(_ context.Context, userID uint) (*service.UserPermissions, error) {
 			return &service.UserPermissions{
 				IsSuperAdmin:    true,
-				TeamPermissions: map[uint][]string{},
+				TeamPermissions: map[int64][]string{},
 			}, nil
 		},
 	}
